Move entity type into repository definition file
