--- src/provider/model/Entity.ts
@@ -1,6 +1,0 @@
-/**
- * @description Base model for our entities
- */
-export interface Entity {
-    id: string;
-}
